Changed Static Factory Method to Non Static Method
Defining a simple factory as a static method is a common technique
and is often called a static factory. Why use a static method? Because you
don’t need to instantiate an object to make use of the create method. But it
also has the disadvantage that you can’t subclass and change the behavior
of the create method.